Add files via upload
R code to create two different pedigrees, one based on the "pedigree" table, the other on the "flastall" table. The latter is nearly twice as large.

Also code to detect mothers/fathers present as dams/sires but not in their own right. These only occur in the flastall based pedigree, 48 dams and 1 sire.